chore(tests): #13 - Prepare dockerfile running unit and integration tests

+++ hack/get-kube-binaries.sh
@@ -20,22 +20,38 @@ ETCD_VERSION="${ETCD_VERSION:-v3.2.18}"
 mkdir -p hack/bin
 cd hack/bin
 
-# Download kubectl.
-rm -f kubectl
-wget  -nv "${KUBERNETES_RELEASE_URL}/${KUBE_VERSION}/bin/linux/amd64/kubectl"
-chmod +x kubectl
+# Download kubectl if not found
+if [[ -f ./kubectl ]] && ./kubectl version --client; then
+    echo ""
+    echo "+++ Above kubectl was installed previously"
+    echo ""
+else
+    wget -nv "${KUBERNETES_RELEASE_URL}/${KUBE_VERSION}/bin/linux/amd64/kubectl"
+    chmod +x kubectl
+fi
 
-# Download kube-apiserver.
-rm -f kube-apiserver
-wget -nv "${KUBERNETES_RELEASE_URL}/${KUBE_VERSION}/bin/linux/amd64/kube-apiserver"
-chmod +x kube-apiserver
 
-# Download etcd.
-rm -f etcd
-basename="etcd-${ETCD_VERSION}-linux-amd64"
-filename="${basename}.tar.gz"
-url="https://github.com/coreos/etcd/releases/download/${ETCD_VERSION}/${filename}"
-wget  -nv "${url}"
-tar -zxf "${filename}"
-mv "${basename}/etcd" etcd
-rm -rf "${basename}" "${filename}"
+# Download kube-apiserver if not found
+if [[ -f ./kube-apiserver ]] && ./kube-apiserver --version; then
+    echo ""
+    echo "+++ Above kube-apiserver was installed previously"
+    echo ""
+else
+    wget -nv "${KUBERNETES_RELEASE_URL}/${KUBE_VERSION}/bin/linux/amd64/kube-apiserver"
+    chmod +x kube-apiserver
+fi
+
+# Download etcd if not found
+if [[ -f ./etcd ]] && ./etcd --version; then
+    echo ""
+    echo "+++ Above etcd was installed previously"
+    echo ""
+else
+    basename="etcd-${ETCD_VERSION}-linux-amd64"
+    filename="${basename}.tar.gz"
+    url="https://github.com/coreos/etcd/releases/download/${ETCD_VERSION}/${filename}"
+    wget -nv "${url}"
+    tar -zxf "${filename}"
+    mv "${basename}/etcd" etcd
+    rm -rf "${basename}" "${filename}"
+fi
